+++ web/src/render/components/QuestionWrapper.vue
@@ -16,6 +16,7 @@ import { useShowInput } from '@/render/hooks/useShowInput'
 import store from '@/render/store'
 import { cloneDeep } from 'lodash-es'
 import { ruleEngine } from '@/render/hooks/useRuleEngine.js'
+import { useQuestionStore } from '../stores/question'
 
 import { NORMAL_CHOICES, RATES, QUESTION_TYPE } from '@/common/typeEnum.ts'
 
@@ -32,6 +33,7 @@ const props = defineProps({
   }
 })
 const emit = defineEmits(['change'])
+const questionStore = useQuestionStore()
 
 const formValues = computed(() => {
   return store.state.formValues
@@ -105,7 +107,7 @@ const handleChange = (data) => {
   emit('change', data)
   // 处理投票题
   if (props.moduleConfig.type === QUESTION_TYPE.VOTE) {
-    store.dispatch('updateVoteData', data)
+    questionStore.updateVoteData(data)
   }
 }
 </script>
