+++ web/src/render/stores/question.js
@@ -1,0 +1,157 @@
+import { ref, computed } from 'vue'
+import { defineStore } from 'pinia'
+import { set } from 'lodash-es'
+import { useSurveyStore } from '@/render/stores/survey'
+import { queryVote } from '@/render/api/survey'
+
+const VOTE_INFO_KEY = 'voteinfo'
+
+export const useQuestionStore = defineStore('question', () => {
+  const voteMap = ref({})
+  const questionData = ref(null)
+  const questionSeq = ref([]) // 题目的顺序，因为可能会有分页的情况，所以是一个二维数组[[qid1, qid2], [qid3,qid4]]
+
+  // 题目列表
+  const renderData = computed(() => {
+    let index = 1
+    return (
+      questionSeq.value &&
+      questionSeq.value.reduce((pre, item) => {
+        const questionArr = []
+
+        item.forEach((questionKey) => {
+          const question = { ...questionData.value[questionKey] }
+          // 开启显示序号
+          if (question.showIndex) {
+            question.indexNumber = index++
+          }
+
+          questionArr.push(question)
+        })
+
+        if (questionArr && questionArr.length) {
+          pre.push(questionArr)
+        }
+
+        return pre
+      }, [])
+    )
+  })
+
+  const setQuestionData = (data) => {
+    questionData.value = data
+  }
+
+  const changeSelectMoreData = (data) => {
+    const { key, value, field } = data
+    set(questionData.value, `${field}.othersValue.${key}`, value)
+  }
+
+  const setQuestionSeq = (data) => {
+    questionSeq.value = data
+  }
+
+  const setVoteMap = (data) => {
+    voteMap.value = data
+  }
+
+  const updateVoteMapByKey = (data) => {
+    const { questionKey, voteKey, voteValue } = data
+    // 兼容为空的情况
+    if (!voteMap.value[questionKey]) {
+      voteMap.value[questionKey] = {}
+    }
+    voteMap.value[questionKey][voteKey] = voteValue
+  }
+
+  //初始化投票题的数据
+  const initVoteData = async () => {
+    const surveyStore = useSurveyStore()
+    const surveyPath = surveyStore.surveyPath
+
+    const fieldList = []
+
+    for (const field in questionData.value) {
+      const { type } = questionData.value[field]
+      if (/vote/.test(type)) {
+        fieldList.push(field)
+      }
+    }
+
+    if (fieldList.length <= 0) {
+      return
+    }
+    try {
+      localStorage.removeItem(VOTE_INFO_KEY)
+      const voteRes = await queryVote({
+        surveyPath,
+        fieldList: fieldList.join(',')
+      })
+
+      if (voteRes.code === 200) {
+        localStorage.setItem(
+          VOTE_INFO_KEY,
+          JSON.stringify({
+            ...voteRes.data
+          })
+        )
+        setVoteMap(voteRes.data)
+      }
+    } catch (error) {
+      console.log(error)
+    }
+  }
+
+  const updateVoteData = (data) => {
+    const { key: questionKey, value: questionVal } = data
+    // 更新前获取接口缓存在localStorage中的数据
+    const localData = localStorage.getItem(VOTE_INFO_KEY)
+    const voteinfo = JSON.parse(localData)
+    const currentQuestion = questionData.value[questionKey]
+    const options = currentQuestion.options
+    const voteTotal = voteinfo?.[questionKey]?.total || 0
+    let totalPayload = {
+      questionKey,
+      voteKey: 'total',
+      voteValue: voteTotal
+    }
+    options.forEach((option) => {
+      const optionhash = option.hash
+      const voteCount = voteinfo?.[questionKey]?.[optionhash] || 0
+      // 如果选中值包含该选项，对应voteCount 和 voteTotal  + 1
+      if (
+        Array.isArray(questionVal) ? questionVal.includes(optionhash) : questionVal === optionhash
+      ) {
+        const countPayload = {
+          questionKey,
+          voteKey: optionhash,
+          voteValue: voteCount + 1
+        }
+        totalPayload.voteValue += 1
+        updateVoteMapByKey(countPayload)
+      } else {
+        const countPayload = {
+          questionKey,
+          voteKey: optionhash,
+          voteValue: voteCount
+        }
+        updateVoteMapByKey(countPayload)
+      }
+      updateVoteMapByKey(totalPayload)
+    })
+  }
+
+  return {
+    voteMap,
+    questionData,
+    questionSeq,
+    renderData,
+    setQuestionData,
+    changeSelectMoreData,
+    setQuestionSeq,
+    setVoteMap,
+    updateVoteMapByKey,
+    initVoteData,
+    updateVoteData
+  }
+})
